Add spinbox to set img zoomlevel, cleanup toolbar
This adds a spinbox to allow the user to set the zoomlevel at which the
image is 100% using a spinbox in the toolbar, as well as cleaning up the
toolbar so that it looks cleaner, and shuffling the tools around in it.

There are now tooltips for the spinboxes so you know what they do.