Load image with imageprovider, display in flickable
This loads the geotiff image using an image provider - which vastly
improves performance, as it does the loading asynchronously.
It still uses a GeoTiffHandler to do the image data reading.

The GeoTiffHandler created from QML reads the geotiff metadata without
reading the image data, and a separate GeoTiffHandler owned by the
GeoTiffImageProvider reads the image data.

This also moves to displaying the image in a flickable, which improves
the scrolling behavior, and adds panning with the mouse.