Render geotiff with transformation (slow)
This now renders the geotiff overlayed over the map with the image
transformed using a QTransform.

The QTransform is generated from 3 different transformation matrices,
translating from (1) GeoTiff pixel coordinates to the Spatial reference
system used by the specific GeoTiff, from (2) the GeoTiff's SRS to the
WGS84 SRS - used by Qt Location (a mercator projection), then from the
WGS84 SRS to the pixel coordinates of the Qt location map viewport.

It is very slow right now, and some caching needs to happen to speed
things up. There's no need to do the full transform of the geotiff on
every update - but that is what is happening right now.