Size tiff item to image, set pos and impl width
This sizes the tiff overlay item to the image instead of making it sized
to fit the entire viewport.

Along with this - the transformed image is sized just big enough for the
transformed pixels, and the position and implicit size of the item is
set to match the transformed image.

Added logic to short-circuit the image transform when the image is
entirely offscreen.