Split image gen out, only gen when needed
Split the generation of the transformed image out into a separate
method, and then only do image gen when needed - have a way to mark
dirty.

Only gen when new source is loaded, and when zoomlevel changes.

Note: this may change because we render the entire transformed image,
including parts that are offscreen. We should render only what is in
the viewport.