Handle pointers through std::unique_ptr
The instances that GDAL returns as pointers do not have good lifetime
management. Automate this by wrapping them in std::unique_ptr.

This will ensure that they are properly deleted when they are done.
Additionally, GDALClose does not need to be explicitly called -- it is
done in the destructor of GDALDataset.